pyproject.toml
[project]
name = "lmdb-dict"
description = 'Full-featured Python dict interface to the LMDB "Lightning" Database.'
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
  { name = "Jesse London", email = "jesselondon@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
  "Topic :: Database",
  "Topic :: Database :: Front-Ends",
  "Topic :: Software Development",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["dict", "lmdb"]
dependencies = [
  "cachetools~=5.3",
  "Dickens~=2.0",
  "lmdb>=1.4",
  "pyyaml>=6",
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/chicago-cdac/lmdb-dict#readme"
Issues = "https://github.com/chicago-cdac/lmdb-dict/issues"
Source = "https://github.com/chicago-cdac/lmdb-dict"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "lmdb_dict/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "ipython",
  "pytest",
  "pytest-cov",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=lmdb_dict test {args}"
no-cov = "cov --no-cov {args}"

[[tool.hatch.envs.test.matrix]]
python = ["38", "39", "310", "311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "lmdb_dict/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
